finished refactoring everything but the event update function; added option for bulk adding attendance and gave it some actual UI as well as caching the date to improve consecutive additions.

+++ members/views.py
@@ -1,18 +1,17 @@
+import time
 from itertools import groupby
 from operator import attrgetter
-import json
-import time
+
 from django.http import JsonResponse
 from django.shortcuts import render, redirect, get_object_or_404
 from authentication.models import Member
-from .models import AttendanceRecord
-
-from events.models import Event
 from django.contrib.auth.decorators import user_passes_test, login_required
+from django.db import transaction
+
+from .models import AttendanceRecord
 from authentication.views import is_officer
 
-# Create your views here.
-
+# Views
 @login_required
 def members_view(request):
     members = Member.objects.all()
@@ -23,32 +22,30 @@ def delete_member(request):
     if request.method == "POST":
         n_num = request.POST.get('n_num')
         member = get_object_or_404(Member, username=n_num)
-        if member:
-            member.delete()
+        if member: member.delete()
 
     return redirect("/members/")
 
 @user_passes_test(is_officer)
 def update_user(request):
-    if request.method == "POST":
-        viewing_nnum = request.POST.get('viewing_nnum')
-        viewing_user = get_object_or_404(Member, username=viewing_nnum)
-        if not viewing_user: return redirect(f"/members/{viewing_nnum}")
+    if request.method != "POST": return
 
-        n_num = request.POST.get('n_num')
-        name = request.POST.get('name')
-        password = request.POST.get('password')
+    viewing_nnum = request.POST.get('viewing_nnum')
+    viewing_user = get_object_or_404(Member, username=viewing_nnum)
+    if not viewing_user: return redirect(f"/members/")
 
-        viewing_user.username = n_num
-        viewing_user.name = name
-        
-        if password and password != "":
-            viewing_user.set_password(password)
+    n_num = request.POST.get('n_num')
+    name = request.POST.get('name')
+    password = request.POST.get('password')
 
-        viewing_user.save()
-        return redirect(f"/members/{n_num}")
+    viewing_user.username = n_num
+    viewing_user.name = name
+    
+    if password and password != "":
+        viewing_user.set_password(password)
 
-        
+    viewing_user.save()
+    return redirect(f"/members/{n_num}")
 
 def view_member(request, n_num):
     member = get_object_or_404(Member, username=n_num)
@@ -63,68 +60,64 @@ def view_member(request, n_num):
     return render(request, "view_member.html", {"member": member, "events": participating_events})
 
 def attendance_view(request):
-    # Update user references for accounts created after attendance scanned
-    null_users = AttendanceRecord.objects.filter(user__isnull=True)
-    for user in null_users:
-        try:
-            user_obj = Member.objects.get(username=user.n_number)
-            user.user = user_obj
-            user.save()
-        except Member.DoesNotExist:
-            pass
-    
-    records = AttendanceRecord.objects.select_related('user').order_by('date')
+    null_user_records = AttendanceRecord.objects.filter(user__isnull=True)
 
-    # Group by date
-    grouped_by_date = {
+    for record in null_user_records:
+        user = Member.objects.filter(username=record.n_number).first()
+        if not user: continue
+
+        record.user = user
+        record.save()
+
+    records = AttendanceRecord.objects.select_related('user').order_by('date')
+    records_grouped_by_date = {
         date: list(group)
         for date, group in groupby(records, key=attrgetter('date'))
     }
-    return render(request, "attendance.html", {"records": grouped_by_date})
+
+    return render(request, "attendance.html", {"records": records_grouped_by_date})
 
 def scan_attendance_record(request):
     return render(request, "scan_attendance.html", {})
 
 def add_attendance_record(request):
+    cached_date = ""
     if request.method == "POST":
-        try:
-            n_num = request.POST.get('n_num').upper()
-            date = request.POST.get('date')
-        except:
-            body = json.loads(request.body)
-            n_num = body["n_num"]
-            date = body["date"]
+        n_num = request.POST.get('n_num').upper()
+        date = request.POST.get('date')
+        cached_date = date
 
-        if not n_num or not date:
-            return render(request, "add_attendance_record.html")
-        
-        try:
-            user_obj = Member.objects.get(username=n_num)
-        except Member.DoesNotExist:
-            user_obj = None
+        if not n_num or not date: return render(request, "add_attendance_record.html")
+        if "," in n_num:
+            n_nums = [(not "N" in num and "N" + num or num) for num in n_num.split(",")]
+            user_objs = Member.objects.filter(username__in=n_nums)
+            mapped_list = {user.username: user for user in user_objs}
+            records = [AttendanceRecord(date=date, n_number=username, user=mapped_list.get(username)) for username in n_nums]
 
-        AttendanceRecord.objects.create(
-            date = date,
-            n_number = n_num,
-            user = user_obj
-        )
+            with transaction.atomic():
+                AttendanceRecord.objects.bulk_create(records)
+        else:
+            if not "N" in n_num: n_num = "N" + n_num
 
-    return render(request, "add_attendance_record.html")
+            user_obj = Member.objects.filter(username=n_num).first() or None
+            AttendanceRecord.objects.create(
+                date = date,
+                n_number = n_num,
+                user = user_obj
+            )
+
+    return render(request, "add_attendance_record.html", {"date": cached_date})
     
 def delete_record(request):
     if request.method != "POST": return redirect("/members/attendance/")
 
     date = request.POST.get("date")
     n_num = request.POST.get("n_num")
-    print(date, n_num)
     
-    record = AttendanceRecord.objects.filter(date=date, n_number=n_num)
-    if record.count() > 1:
-        record = record.last()
+    record = AttendanceRecord.objects.filter(date=date, n_number=n_num).last()
+    if not record: return
 
-    if record:
-        record.delete()
-
+    record.delete()
     cache_buster = int(time.time())
     return JsonResponse({
         "redirect_url": f"/members/attendance/?v={cache_buster}"
